fingerprint: goodix_fps_mmi: Rewrite for performance
Goodix's implementation is quite generic at its core and so,
there was room for improvement without hurting the basic functions.
Debugging and unsupported features are removed as well as the display
notifier and a wakelock as our SoC can receive interrupts from sleep
and respond just fine.

Generally, mutexes and other bogus checks have also been removed
and indents have been fixed, driver should be a lot more readable
now that everything is in one place.

This rewrite can be adapted to work on the majority of goodix sensors
with minor edits for pinctrl or gpio names or new regulators and such.

Change-Id: I44239ae2134da81895a3b146292b0e4f3a5a1acc
Signed-off-by: Jebaitedneko <Jebaitedneko@gmail.com>
Signed-off-by: Cyber Knight <cyberknight755@gmail.com>