usb: mtu3: gadget: implement optional explicit status stage
Implement the mechanism for optional explicit status stage for the MTU3
driver. This allows a function driver to specify what to reply for the
status stage. The functionality for an implicit status stage is
retained.

This patch is heavily based on the MUSB version by
"Paul Elder <paul.elder@ideasonboard.com>" together with the ack cleanup
patch by me, adapted and modified for MTU3. Adds a new state because
MTU3 did not have any relevant state to hook into.

Aditionally, this patch also supports explicit status stage for IN
transfers.

Change-Id: If39c0722e7e9cc80c65ab51d03ac40c556bb61da
Signed-off-by: Cyber Knight <cyberknight755@gmail.com>