usb: gadget: uvc: make API compatible with mainline again
The new proposed API in this patchset never gained traction. However, if
userspace decides to, it's still possible to stall if validation of
setup or data fails. Because there is no real reason to hold data stage
in a control OUT transaction, this patch delivers both setup and data
event at once.
Another reason is that delivering setup earlier causes races.

Change-Id: If378c9b977661ae9bd11e10d46e23b6d7bd2afe5
Signed-off-by: Cyber Knight <cyberknight755@gmail.com>